Move expression from common to data package to preven cyclin' imports
